#401 Sisimai::String.token() has been moved to Sisimai::Fact.token()
