+++ lib/sisimai/string.rb
@@ -7,21 +7,6 @@ module Sisimai
         body: %r|<head>.+</head>.*<body[ >].+</body>|im,
       }
 
-      # Create message token from addresser and recipient
-      # @param  [String]  addr1 Sender address
-      # @param  [String]  addr2 Recipient address
-      # @param  [Integer] epoch Machine time of the email bounce
-      # @return [String]        Message token(MD5 hex digest) or blank(failed to create token)
-      # @see    http://en.wikipedia.org/wiki/ASCII
-      def token(addr1, addr2, epoch)
-        return "" if addr1.is_a?(::String) == false || addr2.is_a?(::String) == false
-        return "" if addr1.empty? || addr2.empty? || epoch.is_a?(Integer) == false
-
-        # Format: STX(0x02) Sender-Address RS(0x1e) Recipient-Address ETX(0x03)
-        require 'digest/sha1'
-        return Digest::SHA1.hexdigest(sprintf("\x02%s\x1e%s\x1e%d\x03", addr1.downcase, addr2.downcase, epoch))
-      end
-
       # The argument is 8-bit text or not
       # @param    [String] argvs  Any string to be checked
       # @return   [Boolean]       false: ASCII Characters only, true: Including 8-bit character
